QQplot for rare variants + HWE threshold
The QQplot.R script now accepts the name of the pvalue column as input, so it can now be used for files other than the single variant output from SAIGE.

Also changed the QQplot NF process to account for this change.

Named colors for the het. plot (to avoid color shift when only one or two classes are present instead of the three).

Added p-value threshold for HWE exact test